update file names and explanations
