+++ lib/11c_dog.rb
@@ -1,0 +1,31 @@
+# frozen_string_literal: true
+
+require_relative '../lib/11a_pet'
+
+# This example has multiple files:
+# lib/11a_pet.rb is the parent 'Pet' class of 'Cat' and 'Dog'
+# lib/11b_cat.rb is a subclass 'Cat'
+# lib/11c_dog.rb is a subclass 'Dog'
+# spec/11a_shared_example_spec.rb has shared tests for 'Cat' and 'Dog'
+# spec/11b_cat_spec.rb has the tests for subclass 'Cat'
+# spec/11c_dog_spec.rb has the tests for subclass 'Dog'
+
+# The file order to complete this exercise:
+# 1. Familarize yourself with the 3 files in the lib folder.
+# 2. Review the tests in spec/11a_shared_example_spec.rb
+# 3. Complete either spec/11b_cat_spec.rb or spec/11c_dog_spec.rb
+
+# Dog is a sub-class of Pet
+class Dog < Pet
+  def talk
+    'WOOF!'
+  end
+
+  def sleeping?
+    true
+  end
+
+  def barking?
+    false
+  end
+end
