+++ spec/11b_cat_spec.rb
@@ -1,20 +1,27 @@
 # frozen_string_literal: true
 
-require_relative '../lib/11_cat'
-require_relative '../spec/11_shared_example_spec'
+require_relative '../lib/11b_cat'
+require_relative '../spec/11a_shared_example_spec'
 
-# This example has a 'Pet' class with two subclasses 'Cat' and 'Dog'.
-# There is a spec file for cat & dog (each file begins with '11_').
+# This example has multiple files:
+# lib/11a_pet.rb is the parent 'Pet' class of 'Cat' and 'Dog'
+# lib/11b_cat.rb is a subclass 'Cat'
+# lib/11c_dog.rb is a subclass 'Dog'
+# spec/11a_shared_example_spec.rb has shared tests for 'Cat' and 'Dog'
+# spec/11b_cat_spec.rb has the tests for subclass 'Cat'
+# spec/11c_dog_spec.rb has the tests for subclass 'Dog'
 
-# In addition, there is a third spec file called '11_shared_example'.
-# That file has two tests that are used in both the cat & dog spec files.
+# The file order to complete this exercise:
+# 1. Familarize yourself with the 3 files in the lib folder.
+# 2. Review the tests in spec/11a_shared_example_spec.rb
+# 3. Complete either spec/11b_cat_spec.rb or spec/11c_dog_spec.rb
 
 describe Cat do
   subject(:oscar) { described_class.new('Oscar', 'Maine Coon') }
 
-  # Check out the file 11_shared_examples_spec.rb.
-  # This test references that file's first test in the 'include_examples' line.
-  # The shared_example_spec file has further information.
+  # Before you begin this file, make sure you have read the shared
+  # example file: 11a_shared_examples_spec.rb. This test references
+  # that file's first test in the below 'include_examples' line.
   context 'when Cat is a child class of Pet' do
     include_examples 'base class method name'
   end
@@ -28,7 +35,7 @@ describe Cat do
 end
 
 # ASSIGNMENT - complete either Cat or Dog assignment
-# (see 11_dog_spec.rb for Dog assignment)
+# (see 11c_dog_spec.rb for Dog assignment)
 
 describe Cat do
   # Create a subject with your choice of cat name and optional breed/color.
